LayoutTests/fast/dom/Element/id-in-applet.html

<html>
<head>
<link rel="stylesheet" href="../../js/resources/js-test-style.css">
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>

<applet id="applet1" width="10" height="10"></applet>

<script>
description("Test using id in applet element");

shouldBeNonNull('document.getElementById("applet1")');

var successfullyParsed = true;
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
