Safari-609.1.11.1/LayoutTests/fast/dom/Element/id-in-applet-expected.txt

Test using id in applet element

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById("applet1") is non-null.
PASS successfullyParsed is true

TEST COMPLETE
